hide chunkers and hashers behind features
