+++ runner/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "runner"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+chunkfs = { path = "../.", features = ["chunkers", "hashers"] }
+rand = "0.8.5"
